Changed float to double

+++ include/ibm.h
@@ -8,8 +8,8 @@
 
 	void interpolation(fluid fluido, mesh membrana, int x, int y, int z);
 	void spread(fluid fluido, mesh membrana, int x, int y, int z);
-	double dirac_2(double *x);
-	double dirac_3(double *x);
-	double dirac_4(double *x);
+	float dirac_2(float *x);
+	float dirac_3(float *x);
+	float dirac_4(float *x);
 
 #endif
